chore: slight reorg for consistency
